Move bitcoind out of utils and fix some obvious things
